refactor(bezier): rename the kernel header and move both into pantr::bezier
cpp/include/pantr/bezier/bezier.hpp becomes kernels_1d.hpp, because this
directory is about to gain a Bezier type and two headers called bezier
side by side is a trap: an include line would no longer say which one a
translation unit meant.

Both bezier headers also move from namespace pantr into pantr::bezier,
which is what pantr::grid, pantr::geometry and pantr::transform already
do. root_finding.hpp's internal `detail` namespace moves with it, which
also stops it sharing pantr::detail with quad/legendre.hpp.

bezier/bezier.hpp stays as a one-line forwarder. The top-level
CMakeLists.txt installs cpp/include/pantr wholesale and exports a
findable package with COMPATIBILITY SameMinorVersion, so the installed
header set is a promise. It forwards the PATH, not the old spelling of
the names -- an installed consumer that included it still compiles, and
still has to say pantr::bezier:: to call anything.

Beyond the five dependents this touches two the rename would otherwise
have broken silently: scripts/measure_bezier_fma_bound.py embeds a C++
translation unit that calls the kernels by qualified name, and
cpp/consumer/main.cpp now includes the forwarder deliberately. The
consumer is the one program compiled against the INSTALLED package, so it
is the only build that would notice the forwarder being deleted; it also
calls one kernel through it, so the include is not merely parsed.
